+++ apps/species/forms.py
@@ -50,26 +50,15 @@ class SpeciesFilter(FilterForm):
         ),
     )
 
-    family = forms.ChoiceField(
-        label=_("Family"),
-        required=False,
-        widget=autocomplete.Select2(
-            url="species:families",
-            attrs={"class": "form-select", "data-theme": "bootstrap-5"},
-        ),
-    )
-
     filters = {
         "common_name": "species",
         "scientific_name": "species",
-        "family": "species__family_code",
     }
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         if self.is_bound:
             self.fields["common_name"].choices = self.get_common_name_choice()
-            self.fields["family"].choices = self.get_family_choice()
 
     def get_common_name_choice(self):
         choices = []
@@ -82,71 +71,3 @@ class SpeciesFilter(FilterForm):
             if choice:
                 choices = [(choice[0], json.loads(choice[1])[get_language()])]
         return choices
-
-    def get_family_choice(self):
-        choices = []
-        if code := self.data.get("family"):
-            choice = (
-                Species.objects.filter(family_code=code)
-                .values_list("family_code", "family_scientific_name")
-                .first()
-            )
-            if choice:
-                choices = [choice]
-        return choices
-
-
-class FamilyFilter(FilterForm):
-    form_id = "family"
-    form_title = _("By Family")
-
-    family = forms.ChoiceField(
-        label=_("Family"),
-        required=False,
-        widget=autocomplete.Select2(
-            url="species:families",
-            attrs={"class": "form-select", "data-theme": "bootstrap-5"},
-        ),
-    )
-
-    filters = {
-        "family": "species__family_code",
-    }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if self.is_bound:
-            self.fields["family"].choices = self.get_family_choice()
-
-    def get_family_choice(self):
-        choices = []
-        if code := self.data.get("family"):
-            choice = (
-                Species.objects.filter(family_code=code)
-                .values_list("family_code", "family_scientific_name")
-                .first()
-            )
-            if choice:
-                choices = [choice]
-        return choices
-
-
-class SpeciesOrder(FilterForm):
-    form_id = "species-order"
-    form_title = _("Order By")
-
-    order = forms.ChoiceField(
-        label=_("Ordering"),
-        choices=(
-            ("species,started", _("First Seen")),
-            ("species,-started", _("Last Seen")),
-            ("species,-count", _("Highest Count")),
-        ),
-        required=False,
-        widget=forms.Select(attrs={"class": "form-control"}),
-    )
-
-    def get_ordering(self):
-        if order := self.cleaned_data.get("order"):
-            return order.split(",")
-        return "species", "started"