+++ apps/species/urls.py
@@ -4,7 +4,6 @@ from .views import SpeciesView
 
 from .views import (
     CommonNameList,
-    FamilyList,
     ScientificNameList
 )
 
@@ -14,5 +13,4 @@ urlpatterns = [
     path("", SpeciesView.as_view(), name="list"),
     path("common-names", CommonNameList.as_view(), name="common-names"),
     path("scientific-names", ScientificNameList.as_view(), name="scientific-names"),
-    path("families", FamilyList.as_view(), name="families"),
 ]